+++ src/main/java/br/diabetes/permissao/PermissaoService.java
@@ -1,43 +1,55 @@
 package br.diabetes.permissao;
 
-import java.util.List;
 import java.util.Optional;
-
+import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
+import br.diabetes.permissao.Permissao;
+import br.diabetes.permissao.comandos.BuscarPermissao;
+import br.diabetes.permissao.PermissaoId;
 import br.diabetes.permissao.comandos.CriarPermissao;
 import br.diabetes.permissao.comandos.EditarPermissao;
+import br.diabetes.usuario.UsuarioId;
 
 @Service
 @Transactional
 public class PermissaoService {
 	@Autowired
 	private PermissaoRepository repo;
-	
-	public Optional<PermissaoId> executar(CriarPermissao comando){
-		Permissao novo = repo.save(new Permissao(comando));
-		return Optional.of(novo.getId());
+
+	public Optional<PermissaoId> salvar(CriarPermissao comando, UsuarioId id) {
+		if (comando.getEmail() != null) {
+			Permissao nova = repo.save(new Permissao(comando, id));
+			return Optional.of(nova.getId());
+		}
+		return Optional.empty();
 	}
 	
-	public List<Permissao> encontrarTodos() {
-		return repo.findAll();
+	public Optional<BuscarPermissao> encontrar(UsuarioId id) {
+		Permissao permissao = repo.findAll(id.toString());
+		if (permissao != null) {
+			return Optional.of(new BuscarPermissao(permissao));
+		}
+		return Optional.empty();
 	}
 	
-	public Optional<Permissao> encontrar(PermissaoId id) {
-		return repo.findById(id);
+	public Optional<PermissaoId> alterar(EditarPermissao comando, UsuarioId usuarioId) {
+		Permissao optional = repo.findById(comando.getId().toString(), usuarioId.toString());
+		if (optional != null) {
+			Permissao per = optional;
+			per.apply(comando);
+			repo.save(per);
+			return Optional.of(comando.getId());
+		}
+		return Optional.empty();
 	}
-	
-	public void deletar(PermissaoId id) {
-		repo.deleteById(id);
-	}
-	
-	public Optional<PermissaoId> alterar(EditarPermissao comando) {
-		Optional<Permissao> optional = repo.findById(comando.getId());
-		Permissao Permissao = optional.get();
-		Permissao.apply(comando);
-		repo.save(Permissao);
-		return Optional.of(comando.getId());
+
+	public Optional<String> deletar(PermissaoId id, UsuarioId usuarioId) {
+		if (repo.findById(id.toString(), usuarioId.toString()) != null) {
+			repo.deleteById(id);
+			return Optional.of("Permissão ===> " + id + ": deletada com sucesso");
+		}
+		return Optional.empty();
 	}
 }
